+++ Managers/Interfaces/IOrderManager.cs
@@ -6,7 +6,7 @@ namespace BuyMore.Managers.Interfaces
 {
     public interface IOrderManager
     {
-        Order CreateOrder(User user, Cart cart, double totalAmount);
+        Order CreateOrder(User user, Cart cart, decimal totalAmount);
         Order? GetOrderByReference(string reference);
         IEnumerable<Order> GetOrdersByUser(int userId);
         IEnumerable<Order> GetAllOrders();
